requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
scikit-learn
altair